+++ src/util/standard/index.ts
@@ -2,5 +2,5 @@
 export * from "./opaque";
 export { Primitive } from "./primitive";
 export * from "./to-primitive";
-export * from "./type";
+export * from "./types";
 export * from "./type-of";
